Refactor drop_tables test case to share more code
