+++ tests/test_duckdb.py
@@ -1,0 +1,26 @@
+from unittest import mock
+from unittest.mock import MagicMock
+
+import duckdb
+from pytest_cases import parametrize_with_cases
+
+from .test_cases import drop_tables_cases as dtc
+
+
+@mock.patch("ldlite.request_get")
+@parametrize_with_cases("tc", cases=dtc.DropTablesCases)
+def test_drop_tables(request_get_mock: MagicMock, tc: dtc.DropTablesCase) -> None:
+    from ldlite import LDLite as uut
+
+    ld = uut()
+    tc.patch_request_get(ld, request_get_mock)
+    dsn = f":memory:{tc.db}"
+    ld.connect_db(dsn)
+
+    for prefix in tc.values:
+        ld.query(table=prefix, path="/patched")
+    ld.drop_tables(tc.drop)
+
+    with duckdb.connect(dsn) as res:
+        res.execute("SHOW TABLES;")
+        assert sorted([r[0] for r in res.fetchall()]) == sorted(tc.expected_values)
